@@ -4,7 +4,12 @@ package com.example.julius.mp3_soitin.views;
  * Created by Julius on 19.2.2018.
  */
 
-public interface BasePresenter {
+public interface BasePresenter<E> {
     void start();
     void stop();
+    void setContentByHelperObject(E e);
+    void setContentByID(int id);
+    void refresh();
+    boolean isActive();
+    boolean onBackPressed();
 }
