@@ -1,6 +1,7 @@
 package com.example.julius.mp3_soitin.views.album;
 
 import com.example.julius.mp3_soitin.data.entities.Album;
+import com.example.julius.mp3_soitin.data.entities.ArtistWithAlbums;
 import com.example.julius.mp3_soitin.views.BasePresenter;
 
 import java.util.List;
@@ -15,7 +16,7 @@ public interface AlbumContract {
         void setPresenter(AlbumContract.Presenter presenter);
     }
 
-    interface Presenter extends BasePresenter {
+    interface Presenter extends BasePresenter<ArtistWithAlbums> {
         void selectAlbum(Album album);
     }
 }
