@@ -14,8 +14,11 @@ import android.widget.ListView;
 import com.example.julius.mp3_soitin.R;
 import com.example.julius.mp3_soitin.filescanning.BadFile;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.CompletableFuture;
+import java.util.function.Consumer;
 
 /**
  * Created by Julius on 26.3.2018.
@@ -28,11 +31,15 @@ public class SimpleListDialog extends DialogFragment {
     /**
      * @return
      */
-    public static SimpleListDialog newInstance(ArrayList<BadFile> list) {
+    public static SimpleListDialog newInstance(ArrayList<BadFile> list, Serializable callback) {
         SimpleListDialog listDialog = new SimpleListDialog();
         if(list!=null){
             Bundle args = new Bundle();
             args.putParcelableArrayList("list", list);
+            //Consumer<Object> o = (oi) -> {};
+            Serializable r = (Consumer<Object> & Serializable)callback;
+            //Serializable c = (Serializable & Consumer<Object>);
+            args.putSerializable("consumer", callback);
             listDialog.setArguments(args);
         }
         return listDialog;
@@ -42,8 +49,9 @@ public class SimpleListDialog extends DialogFragment {
     @NonNull
     public Dialog onCreateDialog(Bundle savedInstanceState) {
         List list = getArguments().getParcelableArrayList("list");
+        Consumer<List<BadFile>> consumer = (Consumer<List<BadFile>>) getArguments().getSerializable("consumer");
         if(arrayAdapter==null) {
-            arrayAdapter = new ArrayAdapter<BadFile>(getContext(), android.R.layout.select_dialog_singlechoice, this.lista);
+            arrayAdapter = new ArrayAdapter<>(getContext(), android.R.layout.select_dialog_singlechoice, this.lista);
             lista.addAll(list);
             arrayAdapter.notifyDataSetChanged();
         }
@@ -54,6 +62,11 @@ public class SimpleListDialog extends DialogFragment {
         builder.setView(convertView);
         ListView lv = convertView.findViewById(R.id.listViewBadFiles);
         lv.setAdapter(arrayAdapter);
+        // Add action buttons
+        builder.setPositiveButton("Tallenna vioista huolimatta", (dialog, id) -> {
+            CompletableFuture.supplyAsync(() -> lista).thenAccept(consumer);
+        })
+        .setNegativeButton("Älä tallenna", (dialog, id) -> SimpleListDialog.this.getDialog().cancel());
         // Inflate and set the layout for the dialog
         // Pass null as the parent view because its going in the dialog layout
         return builder.create();
