@@ -1,7 +1,9 @@
 package com.example.julius.mp3_soitin.views.track;
 
 import com.example.julius.mp3_soitin.data.entities.Track;
+import com.example.julius.mp3_soitin.data.entities.TrackContainer;
 import com.example.julius.mp3_soitin.views.BasePresenter;
+import com.example.julius.mp3_soitin.views.BaseView;
 import com.example.julius.mp3_soitin.views.dialogs.DialogTrackPlaylistUseCase;
 
 import java.util.List;
@@ -12,17 +14,21 @@ import java.util.List;
 
 public interface TrackContract {
 
-    interface View<P>{
+    interface View {
         void showTracks(List<Track> tracks);
         void setWindowName(String newName);
         void resetWindowName();
-        void setPresenter(P presenter);
+        void setPresenter(Presenter presenter);
         void showDialog(DialogTrackPlaylistUseCase usecase, Track track, int mode);
     }
 
-    interface Presenter extends BasePresenter{
+    interface Presenter extends BasePresenter<TrackContainer>{
         void openTrack(int index);
         void longPressedTrack(int index, int mode);
+        void setTracks(List<Track> persistedTracks);
+        void setContainer(TrackContainer container);
+        TrackContainer getContainer();
+        boolean shouldUsePersistedData();
     }
 
 }
